Support power/energy for more Aqara (H1) switches, add devices (#2784)
* Have `OppleCluster` in `opple_remote.py` inherit `XiaomiAqaraE1Cluster`

* Add `MODELS_INFO` to `XiaomiOpple2ButtonSwitchFace1`, `XiaomiOpple2ButtonSwitchFace2`

These are the only models I found online.

* Add power and energy support for `lumi.switch.n2aeu1`, `lumi.switch.b2naus01`

Seems to be supported

* Add alternative signature for `lumi.switch.n1aeu1`, support power/energy, refactor

* Remove redundant `XiaomiMeteringCluster`

* Use `LUMI` constant for `MODELS_INFO`

* Remove cluster ID comments for opple_switch.py

* Remove redundant device_automation_triggers assignment

* Rename `switch_h1.py` to `switch_h1_single.py`, add comment

* Add `switch_h1_double.py` and move H1 double switch quirk from `opple_switch.py`

* Remove cluster ID comments from `XiaomiOpple2ButtonSwitchFace2`

* Rename `lumi.switch.n2aeu1` quirk to `AqaraH1DoubleRockerSwitchWithNeutral`

* Change imports slightly

* Add support for `lumi.switch.l2aeu1`, `lumi.switch.l2aeu1`

* Improve docstring for opple_switch.py
