+++ zhaquirks/xiaomi/aqara/opple_remote.py
@@ -50,7 +50,12 @@ from zhaquirks.const import (
     VALUE,
     ZHA_SEND_EVENT,
 )
-from zhaquirks.xiaomi import LUMI, BasicCluster, XiaomiCustomDevice
+from zhaquirks.xiaomi import (
+    LUMI,
+    BasicCluster,
+    XiaomiAqaraE1Cluster,
+    XiaomiCustomDevice,
+)
 
 PRESS_TYPES = {0: "hold", 1: "single", 2: "double", 3: "triple", 255: "release"}
 STATUS_TYPE_ATTR = 0x0055  # decimal = 85
@@ -91,7 +96,6 @@ COMMAND_6_TRIPLE = "6_triple"
 COMMAND_6_HOLD = "6_hold"
 COMMAND_6_RELEASE = "6_release"
 
-OPPLE_CLUSTER_ID = 0xFCC0
 OPPLE_MFG_CODE = 0x115F
 
 
@@ -129,11 +133,9 @@ class MultistateInputCluster(CustomCluster, MultistateInput):
             super()._update_attribute(0, action)
 
 
-class OppleCluster(CustomCluster):
+class OppleCluster(XiaomiAqaraE1Cluster):
     """Opple cluster."""
 
-    ep_attribute = "opple_cluster"
-    cluster_id = OPPLE_CLUSTER_ID
     attributes = {
         0x0009: ("mode", types.uint8_t, True),
     }
